adapted new fcns/tests after rebasing
