+++ go/cs/reservation/BUILD.bazel
@@ -22,6 +22,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//go/lib/colibri/reservation:go_default_library",
+        "//go/lib/util:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
 )
